requirements/base.txt
fastapi==0.61.1
pydantic[dotenv]==1.6.1
boto3==1.15.1
boto3-stubs[essential]==1.15.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.1.0
python-multipart==0.0.5
